Update get-started.md
With launchpad being deprecated, a new public graphql server needs to be created on codesandbox or glitch. I created a glitch 'playground' at the end point 'https://olive-cushion.glitch.me/graphql'. If you would like to use this endpoint in the documents, or take the node converted code and create an apollo controlled end-point, feel free.